requirements.txt
matplotlib==3.7.2
numpy==1.24.4
onnxruntime==1.15.1
opencv_python==4.8.0.76pi
pycocotools==2.0.7
torchvision==0.13.1+cu113
pyqt5
